add architecture
