+++ parse/category.js
@@ -1,0 +1,12 @@
+/**
+ * Created by artemyel on 15.03.17.
+ */
+var parseCategory = function (data, callback) {
+    var category = [];
+    data.d.ShortName.forEach(function (item, i) {
+        category.push({"ShortName": item, "ID": data.d.ID[i], "ProductCount": data.d.ProductCount[i]})
+    });
+    callback(category);
+};
+
+module.exports = parseCategory;
