+++ parse/product.js
@@ -1,0 +1,10 @@
+/**
+ * Created by artemyel on 15.03.17.
+ */
+var winston = require('winston');
+
+var parseProduct = function (data, callback) {
+
+};
+
+module.exports = parseProduct;
